Got my parser setup but I have an issue with the way the model is setup. I'll fix that later. (#7)
Worked out the kinks in the code. It now returns a Foundation object that will allow me to change data

Removed call to func
updated func name
Made adjustments to the json structure
I got rid of the URLSession instance now that I know I could just create data object based off of URL ;)
Might need some help with this. Pushing it for review.
updated url to file to be immutable
Added a check for if the file is nil. Now it will only add the entry if the file is empty.
Updated json to use array for storing all data.
Still having an issue with adding new entry to json file. Working on getting that going.
Quick refactor
Changed entry id to start at 0
Having some issues so I'm starting this function from scratch.
Function now writes to the json file.